Create Lab 2
The waiting list is created interactively with the following commands–1 <name> <number>•Adds the name and number of people specified to the bottom of the list. Don’t worry with duplicates.–2 <size>•Removes (show and delete) oldest entry that fits the sizeof the opening. Move requests up.–3•Shows the list, name and number, from oldest to newest–0•Quits