Create Lab 8
Waiting list, extend lab 7
Add the following options–5 –show each list backwards•Use a recursivefunction–6 –show each name backwards•Use a recursivefunction–0 –save and quit•Change the function that deletes all the nodes to be recursive •The recursive function traverses each list and it needs to be called from a loop that traverses the array. Have that in main or create another function.